requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
anyio==4.9.0
boto3==1.38.28
botocore==1.38.28
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Cython==3.1.1
easydict==1.13
elastic-transport==8.17.1
elasticsearch==8.10.0
fastapi==0.115.12
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.1
fsspec==2025.5.1
h11==0.16.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
insightface==0.7.3
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnx==1.18.0
onnxruntime==1.22.0
onnxruntime-gpu==1.22.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.2.1
prettytable==3.16.0
protobuf==6.31.1
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.13.0
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.9.0
simsimd==6.4.7
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
stringzilla==3.12.5
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.6.1
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
triton==3.3.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
wcwidth==0.2.13
